3rd commit

+++ src/categories/categories.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { Repository } from 'typeorm';
 import { Category } from './entities/category.entity';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Product } from 'src/products/entities/product.entity';
 
 @Injectable()
 export class CategoriesService {
@@ -11,6 +12,8 @@ export class CategoriesService {
   constructor(
     @InjectRepository(Category)
     private categoryRepository: Repository<Category>,
+    @InjectRepository(Product)
+    private productRepository: Repository<Product>,
   ) {}
   async create(categoryData: Partial<Category>): Promise<Category> {
     const category = this.categoryRepository.create(categoryData);
@@ -35,10 +38,28 @@ export class CategoriesService {
   }
 
   async remove(id: number): Promise<void> {
-    const category = await this.categoryRepository.findOneBy({id});
+    // Check if category exists
+    const category = await this.categoryRepository.findOneBy({ id });
     if (!category) {
       throw new NotFoundException(`Category with ID ${id} not found`);
     }
+
+    // Check if the category is used in any product
+    const products = await this.productRepository.find({ where: { category } });
+    if (products.length > 0) {
+      throw new ConflictException('Category cannot be deleted because it is assigned to one or more products');
+    }
+
+    // Remove the category
     await this.categoryRepository.remove(category);
+  }
+
+  async getCategoryByName(name: string): Promise<Category[]> {
+    const category = await this.categoryRepository.find({ where: { name: name} });
+    console.log(category);
+    if (!category.length) {
+      throw new NotFoundException(`No orders found with status containing "${name}"`);
+    }
+    return category;
   }
 }
